+++ web/apps/photos/src/components/Collections/CollectionShare/publicShare/EnablePublicShareOptions.tsx
@@ -2,7 +2,7 @@ import {
     MenuItemDivider,
     MenuItemGroup,
     MenuSectionTitle,
-} from "@/new/common/components/Menu";
+} from "@/new/shared/components/Menu";
 import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
 import DownloadSharp from "@mui/icons-material/DownloadSharp";
 import LinkIcon from "@mui/icons-material/Link";
