+++ web/apps/photos/src/components/Collections/CollectionShare/publicShare/manage/index.tsx
@@ -1,4 +1,4 @@
-import { MenuItemDivider, MenuItemGroup } from "@/new/common/components/Menu";
+import { MenuItemDivider, MenuItemGroup } from "@/new/shared/components/Menu";
 import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
 import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import RemoveCircleOutline from "@mui/icons-material/RemoveCircleOutline";
