+++ web/apps/photos/src/components/Collections/CollectionShare/sharingDetails.tsx
@@ -2,7 +2,7 @@ import {
     MenuItemDivider,
     MenuItemGroup,
     MenuSectionTitle,
-} from "@/new/common/components/Menu";
+} from "@/new/shared/components/Menu";
 import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
 import AdminPanelSettingsIcon from "@mui/icons-material/AdminPanelSettings";
 import ModeEditIcon from "@mui/icons-material/ModeEdit";
