+++ web/apps/photos/src/components/PhotoViewer/ImageEditorOverlay/ColoursMenu.tsx
@@ -1,4 +1,4 @@
-import { MenuItemGroup, MenuSectionTitle } from "@/new/common/components/Menu";
+import { MenuItemGroup, MenuSectionTitle } from "@/new/shared/components/Menu";
 import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
 import { Box, Slider } from "@mui/material";
 import { t } from "i18next";
